+++ src/main/java/org/anson/miniProject/core/repository/sys/permission/RoleResourceRep.java
@@ -1,17 +1,21 @@
 package org.anson.miniProject.core.repository.sys.permission;
 
+import cn.hutool.core.collection.IterUtil;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import org.anson.miniProject.core.mapper.sys.permission.RoleResourceMapper;
 import org.anson.miniProject.core.model.po.sys.permission.RoleResource;
 import org.anson.miniProject.core.repository.BaseRep;
 import org.anson.miniProject.tool.helper.IdHelper;
 import org.anson.miniProject.tool.helper.InputParamHelper;
+import org.anson.miniProject.tool.helper.LogicDelHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.io.Serializable;
-import java.util.Collection;
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 
 @Component
 @Transactional(rollbackFor = Exception.class)
@@ -34,11 +38,54 @@ public class RoleResourceRep extends BaseRep<RoleResource, RoleResourceMapper> {
         return po.getId();
     }
 
-    public void del(String id){
+    public void del(String id, String operUserId, Date operTime) throws JsonProcessingException {
+        RoleResource po = this.mapper.selectById(id);
+
+        if (po == null){
+            return;
+        }
+
+        this.delHelper.recordDelData(po, operUserId, operTime);
         this.mapper.deleteById(id);
     }
 
-    public void del(Collection<? extends Serializable> idList){
+    public void delByRole(String roleId, String operUserId, Date operTime) throws JsonProcessingException {
+        QueryWrapper<RoleResource> qw = new QueryWrapper<>();
+        qw.eq(RoleResource.ROLEID, roleId);
+
+        List<RoleResource> poList = this.mapper.selectList(qw);
+
+        if (IterUtil.isEmpty(poList)){
+            return;
+        }
+
+        List<String> idList = new ArrayList<>();
+
+        for (RoleResource po : poList){
+            this.delHelper.recordDelData(po, operUserId, operTime);
+            idList.add(po.getId());
+        }
+
+        this.mapper.deleteBatchIds(idList);
+    }
+
+    public void delByResource(String resourceId, String operUserId, Date operTime) throws JsonProcessingException {
+        QueryWrapper<RoleResource> qw = new QueryWrapper<>();
+        qw.eq(RoleResource.RESOURCEID, resourceId);
+
+        List<RoleResource> poList = this.mapper.selectList(qw);
+
+        if (IterUtil.isEmpty(poList)){
+            return;
+        }
+
+        List<String> idList = new ArrayList<>();
+
+        for (RoleResource po : poList){
+            this.delHelper.recordDelData(po, operUserId, operTime);
+            idList.add(po.getId());
+        }
+
         this.mapper.deleteBatchIds(idList);
     }
 
@@ -49,4 +96,6 @@ public class RoleResourceRep extends BaseRep<RoleResource, RoleResourceMapper> {
     public void setMapper(RoleResourceMapper mapper){
         this.mapper = mapper;
     }
+    @Autowired
+    private LogicDelHelper delHelper;
 }
